Move Hotkeys and Theme into app globals
