+++ ui/App/Theme.svelte
@@ -1,6 +1,6 @@
 <script>
-  import { settings } from "./src/session.ts";
-  import { Theme } from "./src/settings.ts";
+  import { settings } from "../src/session.ts";
+  import { Theme } from "../src/settings.ts";
 
   $: {
     const theme = $settings ? $settings.appearance.theme : Theme.Light;
